refactor(package-manager): simplify workspace root detection with `resolve-workspace-root`
