+++ packages/@expo/package-manager/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - Update `npm-package-arg@^7` to `npm-package-arg@^11`. ([#30842](https://github.com/expo/expo/pull/30842) by [@kitten](https://github.com/kitten))
+- Simplify workspace root detection with `resolve-workspace-root`.
 
 ## 1.5.2 — 2024-04-23
